+++ src/App.js
@@ -1,49 +1,27 @@
-import "./App.css";
 import { Component } from "react";
-import { PostCard } from "./components/PostCard";
+
+import "./App.css";
+
+import { Posts } from "./components/Posts";
+import  { loadPosts } from './utils/load-posts'
 
 // Class Components //
-
 class App extends Component {
   state = {
     posts: [],
   };
 
   // Fazendo uma requisição de API externa
-
-  componentDidMount() {
-    this.loadPosts();
+  async componentDidMount() {
+    await this.loadPosts();
   }
 
   loadPosts = async () => {
-    const postResponse = fetch("https://jsonplaceholder.typicode.com/posts");
-    const photoResponse = fetch("https://jsonplaceholder.typicode.com/photos");
-
-    const [posts, photos] = await Promise.all([postResponse, photoResponse]);
-
-    const postsJson = await posts.json();
-    const photosJson = await photos.json();
-
-    const postsAndPhotos = postsJson.map((post, index) => {
-      return { ...post, cover: photosJson[index].url };
-    });
-
+    const postsAndPhotos = await loadPosts();
     this.setState({ posts: postsAndPhotos });
   };
 
   // Termina aqui o ciclo de vida dos componentes
-
-  handleTimeout = () => {
-    const { posts, counter } = this.state;
-    posts[0].title = "O titulo mudou";
-
-    this.timeoutUpdate = setTimeout(() => {
-      this.setState({
-        posts,
-        counter: counter + 1,
-      });
-    }, 1000);
-  };
 
   // Componente refatorado //
 
@@ -52,18 +30,7 @@ class App extends Component {
 
     return (
       <section className="container">
-        <div className="posts">
-          {/*<h1> {counter} </h1>{" "}*/}
-          {posts.map((post) => (
-            <PostCard
-              key={post.id}
-              title={post.title}
-              body={post.body}
-              id={post.id}
-              cover={post.cover}
-            />
-          ))}{" "}
-        </div>
+        <Posts posts={posts} />
       </section>
     );
   }
