+++ src/components/PostCard/index.jsx
@@ -1,10 +1,12 @@
-// Componente refatorado //
+export const PostCard = ({ title, cover, body, id}) => (
+        <div className="post">
+              <img src={cover} alt={title} />
+              <div key={id} className="post-content">
+                <h1> {title} </h1> <p> {body}</p>
+              </div>
+            </div>
+    );
 
-export const PostCard = ({ title, cover, body, id }) => (
-  <div className="post">
-    <img src={cover} alt={title} />
-    <div className="post-content">
-      <h1> {title} </h1> <h1>{body}</h1>
-    </div>
-  </div>
-);
+
+
+
