--- src/components/PostCards/index.jsx
@@ -1,14 +1,0 @@
-export const PostCard = (props) => {
-    console.log(props)
-    return (
-        <h1>Salve seu Zé</h1>
-
-        
-        //<div className="post">
-        //      <img src={post.cover} alt={post.tile} />
-        //      <div key={post.id} className="post-content">
-        //        <h1> {post.title} </h1> <p> {post.body}</p>
-        //      </div>
-        //    </div>
-    );
-}
