+++ src/components/Posts/index.jsx
@@ -1,0 +1,17 @@
+import { PostCard } from "../PostCard"
+
+export const Posts = ({ posts }) => { 
+    return (
+      <div className="posts">
+        {/*<h1> {counter} </h1>{" "}*/}
+        {posts.map((post) => (
+          <PostCard
+            title={post.title}
+            body={post.body}
+            id={post.id}
+            cover={post.cover}
+          />
+        ))}
+      </div>
+    );        
+}
